+++ tools/flat-storage/src/commands.rs
