+++ tools/flat-storage/src/lib.rs
@@ -1,2 +1,2 @@
 #[cfg(feature = "protocol_feature_flat_state")]
-pub mod flat;
+pub mod commands;
